maybe now the test will run first?
