--- tests/end2end/features/test_a_me_first.py
@@ -1,11 +1,0 @@
-import time
-
-import PyQt5.QtWidgets
-
-
-def test_this(qapp):
-    button = PyQt5.QtWidgets.QPushButton()
-    before = time.monotonic()
-    button.show()
-    after = time.monotonic()
-    assert False, '+++++ button.show() runtime: {}'.format(after - before)
