+++ tests/end2end/features/test_backforward_bdd.py
@@ -17,5 +17,18 @@
 # You should have received a copy of the GNU General Public License
 # along with qutebrowser.  If not, see <http://www.gnu.org/licenses/>.
 
+import time
+
+import PyQt5.QtWidgets
+
+
+def test_this(qapp):
+    button = PyQt5.QtWidgets.QPushButton()
+    before = time.monotonic()
+    button.show()
+    after = time.monotonic()
+    assert False, '+++++ button.show() runtime: {}'.format(after - before)
+
+
 import pytest_bdd as bdd
 bdd.scenarios('backforward.feature')
